# backend/docs/swagger.yaml
basePath: /v1
definitions:
...
  dto.ChannelPreferenceRequest:
    properties:
      allowDigest:
        type: boolean
      allowInstant:
        type: boolean
      digestFrequency:
        type: string
      isEnabled:
        type: boolean
      quietHours:
        $ref: '#/definitions/models.QuietHoursConfig'
    type: object
...
  dto.ErrorBody:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/dto.ErrorBody'
      requestId:
        type: string
    type: object
...
  dto.PreferenceResponse:
    properties:
      allowDigest:
        type: boolean
      allowInstant:
        type: boolean
      categorySettings:
        additionalProperties:
          type: boolean
        type: object
      digestFrequency:
        type: string
      id:
        type: string
      isEnabled:
        type: boolean
      quietHours:
        $ref: '#/definitions/models.QuietHoursConfig'
      type:
        $ref: '#/definitions/models.NotificationType'
      userId:
        type: string
    type: object
...
  dto.UpdatePreferenceRequest:
    properties:
      allowDigest:
        type: boolean
      allowInstant:
        type: boolean
      categorySettings:
        additionalProperties:
          type: boolean
        type: object
      digestFrequency:
        type: string
      isEnabled:
        type: boolean
      quietHours:
        $ref: '#/definitions/models.QuietHoursConfig'
      type:
        $ref: '#/definitions/models.NotificationType'
    required:
    - type
    type: object
...
  models.NotificationType:
    enum:
    - sms
    - email
    - push
    - in_app
    - webhook
    - security
    type: string
    x-enum-varnames:
    - NotificationTypeSMS
    - NotificationTypeEmail
    - NotificationTypePush
    - NotificationTypeInApp
    - NotificationTypeWebhook
    - NotificationTypeSecurity
  models.QuietHoursConfig:
    properties:
      end:
        type: string
      start:
        type: string
      timezone:
        type: string
    type: object
host: 127.0.0.1:9002
info:
  contact:
    email: support@minisource.com
    name: API Support
  description: Notification Service for Minisource - Handles Email, SMS, Push, and
    In-App notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Notifier Service API
  version: "2.0"
paths:
...
  /preferences/user/{userId}:
    get:
      consumes:
      - application/json
      description: Retrieve all notification preferences for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Request ID
        in: header
        name: X-Request-Id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PreferenceResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user notification preferences
      tags:
      - Preferences
    put:
      consumes:
      - application/json
      description: Update or create notification preference for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Request ID
        in: header
        name: X-Request-Id
        type: string
      - description: Preference data
        in: body
        name: preference
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PreferenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update notification preference
      tags:
      - Preferences
  /preferences/user/{userId}/category/{category}:
    patch:
      consumes:
      - application/json
      description: Update notification preferences for a specific category
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Notification category (system, alerts, updates, marketing, security)
        in: path
        name: category
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Request ID
        in: header
        name: X-Request-Id
        type: string
      - description: Category preference data
        in: body
        name: preference
        required: true
        schema:
          $ref: '#/definitions/dto.ChannelPreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PreferenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category preference
      tags:
      - Preferences
  /preferences/user/{userId}/channel/{channel}:
    patch:
      consumes:
      - application/json
      description: Update preferences for a specific notification channel (sms, email,
        push, in_app)
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Notification channel (sms, email, push, in_app)
        in: path
        name: channel
        required: true
        type: string
      - description: Tenant ID
        in: header
        name: X-Tenant-Id
        type: string
      - description: Request ID
        in: header
        name: X-Request-Id
        type: string
      - description: Channel preference data
        in: body
        name: preference
        required: true
        schema:
          $ref: '#/definitions/dto.ChannelPreferenceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PreferenceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update channel-specific preference
      tags:
      - Preferences
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"